Implement basic reason -> perform world update
Introduce a new type: World.

World consists of Entities (Dict Id Entity) and a seed, that will be
used to assign Ids to new entities (so they can be uniquely identified).

On every frame world is updated. Entities are mapped to Actions. An
Action represents the intention of an entity. Then actions are
performed, resulting in a new World.

The logic to calculate actuall actions and to perform them is not yet
implemented (every action is Idle and the world is static).

Whole code was put back into one file. Modularization seems premature.