Create snake - like predator
Predators are chasing bugs and when they catch them, turn them into new
Predators and attach to them.

If there is not enough stimuli in the environment a predator will idle.
The more hungry the predator is, the further it will look for stimulea.

The code is a mess to be refactored. Specifically reason of bugs should
be implemented in a way similar to predators.